feat: implement generic media gallery block
